+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -299,6 +299,8 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             navigationBarIndicatorSchemeColor:
                 controller.navBarIndicatorSchemeColor,
             navigationBarIndicatorOpacity: controller.navBarIndicatorOpacity,
+            navigationBarIndicatorRadius:
+                controller.navBarIndicatorBorderRadius,
             navigationBarBackgroundSchemeColor:
                 controller.navBarBackgroundSchemeColor,
             navigationBarOpacity: controller.navBarOpacity,
@@ -321,6 +323,8 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             navigationRailIndicatorSchemeColor:
                 controller.navRailIndicatorSchemeColor,
             navigationRailIndicatorOpacity: controller.navRailIndicatorOpacity,
+            navigationRailIndicatorRadius:
+                controller.navRailIndicatorBorderRadius,
             navigationRailBackgroundSchemeColor:
                 controller.navRailBackgroundSchemeColor,
             navigationRailOpacity: controller.navRailOpacity,